+++ zenaton/traits/zenatonable.py
@@ -1,17 +1,29 @@
 from ..engine import Engine
 from ..query.builder import Builder
+from zenaton.exceptions import InvalidArgumentError
 
 
 class Zenatonable:
     """
-        Sends self as the single job to be executed
+        Sends self as the single job to be dispatched
         to the engine and returns the result
     """
     def dispatch(self):
         return Engine().dispatch([self])
 
     """
-        Sends self as the single job to be dispatched
+        Sends self as the single job to be scheduled
+        to the engine and returns the result
+    """
+    def schedule(self, cron):
+        if not isinstance(cron, str) or cron == "":
+            raise InvalidArgumentError("Param passed to 'schedule' function must be a non empty cron string")
+
+        self.scheduling = {'cron': cron}
+        return Engine().dispatch([self])[0]
+
+    """
+        Sends self as the single job to be executed
         to the engine and returns the result
     """
     def execute(self):