+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added a `intent_id` property when dispatching workflows and tasks, sending events to workflows, and
   pausing/resuming/killing workflows.
 
+- Added scheduling: `schedule(cron)`
+
 ### Changed
 
 ### Deprecated